Update bdd tests for destinations endpoint
